feat: all tools is class
