+++ services/agents/app/core/crews/metrics_analyst/tools.py
@@ -2,13 +2,16 @@ from typing import List
 from crewai.tools import BaseTool
 
 from ..utils import get_tools_with_tracking
-from app.services.ml_models import get_ml_models_info
-from app.services.metrics import get_metrics
+from app.services.ml_models import GetModelDetailsTool
+from app.services.metrics import GetMetricsForModelTool, DoesModelHaveMetricsTool
 
-tools = {
-    "GetMLModelInfo": get_ml_models_info,
-    "GetMLModelMetrics": get_metrics
-}
+_tool_instances = [
+    GetModelDetailsTool(),
+    GetMetricsForModelTool(),
+    DoesModelHaveMetricsTool()
+]
+
+tools = {tool.name: tool for tool in _tool_instances}
 
 def get_tools(
     agent_role:str
