+++ services/agents/app/core/crews/ml_models_searcher/tools.py
@@ -1,15 +1,18 @@
 from typing import List
 from crewai.tools import BaseTool
 
-from app.services.metrics import get_metrics
-from app.services.ml_models import get_ml_models_info, get_all_ml_models_info
+from app.services.metrics import GetMetricsForModelTool, DoesModelHaveMetricsTool
+from app.services.ml_models import GetModelDetailsTool, GetMultipleModelsDetailsTool
 from ..utils import get_tools_with_tracking
 
-tools = {
-    "MLModelALLMetrics": get_all_ml_models_info,
-    "MLModelMetrics": get_metrics,
-    "MLModelInfo": get_ml_models_info
-}
+_tool_instances = [
+    GetMultipleModelsDetailsTool(),
+    GetModelDetailsTool(),
+    DoesModelHaveMetricsTool(),
+    GetMetricsForModelTool()
+]
+
+tools = {tool.name: tool for tool in _tool_instances}
 
 def get_tools(
     agent_role:str
