+++ services/agents/app/core/crews/reporter/tools.py
@@ -1,20 +1,20 @@
 from typing import List
 from crewai.tools import BaseTool
 
-from app.services.agent_history import get_agent_history
-from app.services.metrics import get_metrics
-from app.services.ml_models import get_ml_models_info
 from ..utils import get_tools_with_tracking
+from app.services.agent_history import GetAgentHistoryTool
+from app.services.metrics import GetMetricsForModelTool
+from app.services.ml_models import GetModelDetailsTool
 
-tools = {
-    "AgentHistory": get_agent_history,
-    "MLModelMetrics": get_metrics,
-    "MLModelInfo": get_ml_models_info
-}
+_tool_instances = [
+    GetAgentHistoryTool(),
+    GetMetricsForModelTool(),
+    GetModelDetailsTool(),
+]
 
-def get_tools(
-    agent_role:str
-) -> List[BaseTool]:
+tools = {tool.name: tool for tool in _tool_instances}
+
+def get_tools(agent_role: str) -> List[BaseTool]:
     return get_tools_with_tracking(
         agent_role=agent_role,
         tools=tools
