+++ services/agents/app/services/trainer/__init__.py
@@ -1,210 +1,19 @@
-import requests
-from typing import Dict, Any
-from crewai.tools import tool
+from .utils import health
+from .tools import (
+    GetExampleTrainingConfigTool,
+    GetAllAvailableModelsTool,
+    GetDeviceInfoTool,
+    GetOptimizersTool,
+    GetSchedulersTool,
+    GetMetricsForTrainerTool
+)
 
-from .utils import *
-from app.logs import get_logger
-
-logger = get_logger(__name__)
-
-@tool("GetExampleJSONTrainer")
-@handle_errors
-def get_example_run_config_trainer_json() -> Dict[str, Any]:
-    """НАЗНАЧЕНИЕ: Получить пример JSON конфигурации для запуска тренировки модели.
-    
-    КОГДА ИСПОЛЬЗОВАТЬ:
-    - Когда нужно понять структуру конфигурационного файла для обучения
-    - В качестве шаблона перед созданием своей конфигурации
-    - Когда не уверен в формате входных данных для тренировки
-    
-    ВХОДНЫЕ ДАННЫЕ:
-    - Нет входных параметров
-    
-    ВОЗВРАЩАЕТ:
-    - dict с информацией о настройке конфигурации обучения
-    
-    ПРИМЕР ВЫЗОВА:
-    get_example_run_config_trainer_json()
-    
-    ВАЖНЫЕ ЗАМЕЧАНИЯ:
-    - Используй как отправную точку для создания конфигурации
-    - Не копируй слепо, адаптируй под свою задачу
-    - Пример всегда актуален для текущей версии API
-    """
-    logger.info("🔧 GetExampleJSONTrainer вызван")
-    return get_json(f"/info/example_config")
-
-@tool("GetAllAvailableModels")
-@handle_errors
-def get_type_and_name_models(filter: str) -> Dict[str, Any]:
-    """
-    НАЗНАЧЕНИЕ: Получить список всех доступных архитектур ML моделей с возможностью фильтрации.
-    
-    КОГДА ИСПОЛЬЗОВАТЬ:
-    - Когда нужно узнать, какие модели доступны для обучения
-    - Для поиска конкретной архитектуры (например, все resnet)
-    - Перед выбором модели для своей задачи
-    - Для изучения доступных опций нейросетевых архитектур
-    
-    ВХОДНЫЕ ДАННЫЕ:
-    - filter (str): Фильтр для поиска моделей по названию.
-      Поддерживает wildcard паттерны. ПЕРЕДАВАТЬ ФИЛЬТР В НИЖНЕМ РЕГИСТРЕ!!
-      Примеры:
-      - "*resnet*" - все модели, содержащие "resnet"
-      - "*convnext*" - все модели ConvNeXt
-      - "*efficient*" - все EfficientNet модели
-      - "*" - все доступные модели
-    
-    ВОЗВРАЩАЕТ:
-    - dict с списком моделей:
-    
-    ПРИМЕР ВЫЗОВА:
-    get_type_and_name_models("*resnet*")
-        
-    ВАЖНЫЕ ЗАМЕЧАНИЯ:
-    - Известно, что используется библиотека timm 
-    - Используй фильтр для сужения поиска, чтобы не перегружать ответ
-    - Если filter пустой или "*", вернутся все доступные модели
-    """
-    logger.info(f"🔧 GetAllAvailableModels вызван с фильтром: {filter}")
-    return get_json("/info/ml_models",params={"filter": filter})
-
-@tool("GetInfoDevice")
-@handle_errors
-def get_info_device() -> Dict[str, Any]:
-    """
-    НАЗНАЧЕНИЕ: Получить информацию о технических возможностях оборудования для обучения.
-    
-    КОГДА ИСПОЛЬЗОВАТЬ:
-    - Перед началом обучения, чтобы понять доступные ресурсы
-    - Для выбора оптимального batch_size и модели
-    - Когда нужно узнать, доступна ли GPU
-    - Для планирования времени обучения
-    - При выборе между несколькими конфигурациями обучения
-    
-    ВХОДНЫЕ ДАННЫЕ:
-    - Нет входных параметров
-    
-    ВОЗВРАЩАЕТ:
-    - dict с информацией об оборудовании
-    
-    ПРИМЕР ВЫЗОВА:
-    get_info_device()
-    
-    ВАЖНЫЕ ЗАМЕЧАНИЯ:
-    - Всегда проверяй доступное устройство перед обучением
-    - На CPU обучение будет значительно медленнее
-    - Подбирай batch_size исходя из доступной памяти
-    - Используй recommended_batch_size как ориентир
-    """
-    logger.info("🔧 GetInfoDevice вызван")
-    return get_json("/info/device")
-
-@tool("GetOptimizers")
-@handle_errors
-def get_optimizers() -> Dict[str, Any]:
-    """
-    НАЗНАЧЕНИЕ: Получить список всех доступных оптимизаторов для обучения.
-    
-    КОГДА ИСПОЛЬЗОВАТЬ:
-    - При выборе оптимизатора для обучения модели
-    - Для изучения доступных опций
-
-    ВХОДНЫЕ ДАННЫЕ:
-    - Нет входных параметров
-
-    ВОЗВРАЩАЕТ:
-    - dict со списком оптимизаторов
-
-    ПРИМЕР ВЫЗОВА:
-    get_optimizers()
-    
-    ВАЖНЫЕ ЗАМЕЧАНИЯ:
-    - Все оптимизаторы совместимы с PyTorch и используют ту же конфигурацию
-    - Параметры оптимизатора передаются так же, как в torch.optim
-    - Названия оптимизаторов чувствительны к регистру
-    """
-    logger.info("🔧 GetOptimizers вызван")
-    return get_json("/info/optimizers")
-
-@tool("GetSchedulers")
-@handle_errors
-def get_scheduler() -> Dict[str, Any]:
-    """
-    НАЗНАЧЕНИЕ: Получить список всех доступных планировщиков learning rate.
-    
-    КОГДА ИСПОЛЬЗОВАТЬ:
-    - Для настройки динамического изменения learning rate во время обучения
-    - Когда нужно улучшить сходимость модели
-    - Для изучения доступных стратегий планирования
-    - При выборе между разными планировщиками
-
-    ВХОДНЫЕ ДАННЫЕ:
-    - Нет входных параметров
-
-    ВОЗВРАЩАЕТ:
-    - dict со списком планировщиков
-
-    ПРИМЕР ВЫЗОВА:
-    get_scheduler()
-
-    ВАЖНЫЕ ЗАМЕЧАНИЯ:
-    - Все планировщики совместимы с PyTorch и используют ту же конфигурацию
-    - Параметры передаются так же, как в torch.optim.lr_scheduler
-    - Названия планировщиков чувствительны к регистру
-    """
-    logger.info("🔧 GetScheduler вызван")
-    return get_json("/info/schedulers")
-
-@tool("GetMetricsForTrainer")
-@handle_errors
-def get_metrics() -> Dict[str, Any]:
-    """
-    НАЗНАЧЕНИЕ: Получить список всех доступных метрик для оценки качества модели.
-    
-    КОГДА ИСПОЛЬЗОВАТЬ:
-    - При выборе метрик для оценки качества модели
-    - Для понимания доступных метрик в системе
-    - Перед настройкой конфигурации обучения
-    - Когда нужно выбрать метрику для оптимизации
-    
-    ВХОДНЫЕ ДАННЫЕ:
-    - Нет входных параметров
-    
-    ВОЗВРАЩАЕТ:
-    - dict со списком доступных метрик в формате:
-    
-    ПРИМЕР ВЫЗОВА:
-    get_metrics()
-    
-    ВАЖНЫЕ ЗАМЕЧАНИЯ:
-    - Метрики чувствительны к регистру (используйте точно как в ответе)
-    - loss всегда доступна во время обучения и добавляется автоматически системой тренировок
-    - Рекомендуется использовать несколько метрик для комплексной оценки
-    """
-    logger.info("🔧 GetMetricsForTrainer вызван")
-    return get_json("/info/metrics")
-
-def health() -> dict:
-    try:
-        # Отправляем POST запрос
-        response = requests.get(
-            f"{config_url.TRAINER_URL}/info/health",
-            timeout=30
-        )
-        
-        # Проверяем статус ответа
-        response.raise_for_status()
-        
-        return response.json()
-    
-    except requests.RequestException as e:
-        logger.error(f"Ошибка HTTP при обращении к сервису метрик: {e}")
-        return {
-            "status": "dead"
-        }
-    except Exception as e:
-        logger.error(f"Ошибка при обращении к сервису метрик: {e}")
-        return {
-            "status": "dead"
-        }
+__all__ = [
+    'health',
+    'GetExampleTrainingConfigTool',
+    'GetAllAvailableModelsTool',
+    'GetDeviceInfoTool',
+    'GetOptimizersTool',
+    'GetSchedulersTool',
+    'GetMetricsForTrainerTool'
+]
